Reorganize Result classes and refactor methods in sadu-queries
The Result classes have been restructured and moved to a 'reading' sub-package while the ManipulationQuery classes have been moved to a 'writing' sub-package, for better separation of concerns. Significant refactorings include removal of the DataSource method from SingleResult, adding new comments to various classes for improved understanding, and renaming of classes.
