--- sadu-queries2/src/main/java/de/chojo/sadu/queries/stages/results/Result.java
@@ -1,11 +1,0 @@
-/*
- *     SPDX-License-Identifier: LGPL-3.0-or-later
- *
- *     Copyright (C) RainbowDashLabs and Contributor
- */
-
-package de.chojo.sadu.queries.stages.results;
-
-public interface Result<T> {
-    T result();
-}
