+++ sadu-queries2/src/main/java/de/chojo/sadu/queries/stages/results/reading/MultiResult.java
@@ -4,10 +4,17 @@
  *     Copyright (C) RainbowDashLabs and Contributor
  */
 
-package de.chojo.sadu.queries.stages.results;
+package de.chojo.sadu.queries.stages.results.reading;
 
 import de.chojo.sadu.queries.stages.Query;
 import de.chojo.sadu.queries.stages.base.QueryProvider;
+import de.chojo.sadu.queries.stages.execution.reading.MappedQuery;
+
+/**
+ * Result of a {@link MappedQuery#all()} call
+ *
+ * @param <T> Type of returned object
+ */
 
 public class MultiResult<T> implements QueryProvider, Result<T> {
     private final QueryProvider query;
